requirements.txt
absl-py==2.3.1
asgiref==3.11.1
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coverage==7.13.5
dj-database-url==3.1.2
Django==6.0.3
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==25.3.0
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mysql-connector-python==9.6.0
mysqlclient==2.2.8
namex==0.1.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.3
pytest-django==4.12.0
python-dotenv==1.2.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
